requirements.txt
attrs==22.2.0
jsonschema==4.17.3
mysql-connector-python==8.0.31
protobuf==3.20.1
pyrsistent==0.19.3
